Implement MD5 checksums, refractor code, and more
MainWindow.java
- Implement choosing "StarCraft.exe" file from button
in the UI.
- Implement aesthetic changes to the file drop component
based on if the program is ready to launch bot

BwHeadless.java
- Implement useful constants.
- Implement "String getReadyError()" to determine why
the program is not ready to launch.
- Implement "void dropFile(File file)" for the analysis of dropped
files. E.g. examining whether or not a dropped DLL is a
"BWAPI.dll" or a bot DLL.

FileArray.java
- Implement "boolean isValidIndex(int index)" for index out of
bounds checking.

MainTools.java
- Implement MD5 checksum functionality accessible via
"String getMD5Checksum(String path)".
- Implement UI message box windows.

TokenArray.java
- Implement "boolean isValidIndex(int index)" for index out of
bounds checking.