Implement changes to default config file persistency
- Disable "Local PC" gametype

BwHeadless.java
- Implement "boolean writeDefaultConfigFile()" to call
after a variable is set.

ConfigFile.java
- Change implementation so that the configuration file changes
get written to disk when a variable is changed.

Notes:
- Should now be able to use "BwHeadless" set functions such as
"setBotClient(File file)" and "setBotDll(File file)" to
store their values in memory and on disk. When a variable is set,
it gets written to the default configuration file:
"BwHeadless.DEFAULT_CFG_FILE".