Refractor "BwHeadless.java"
- Remove code that copies a file to the StarCraft directory
if a class member file is set. E.g. "setBwapiDlL()" will no
longer copy the BWAPI.dll file to the StarCraft directory.
Copying of files to the StarCraft directory might be moved to
the invocation of the bot.

Apply some minor changes to "ConfigFile.java" and "MainTools.java"

MainTools.java
- Remove "non-null empty string detected" debugging warning.

ConfigFile.java
- "boolean open(File file)" will now create the file and return true
if the file did not previously exist.